+++ AlertingApp/src/com/vz/usage/util/CustAlert.java
@@ -5,14 +5,14 @@ public class CustAlert {
 	int custid;
 	int typ;
 	String tim;
-	int usage;
+	long usage;
 	
 	@Override
 	public String toString() {
 		return "CustAlert [custid=" + custid + ", typ=" + typ + ", tim=" + tim
 				+ ", usage=" + usage + "]";
 	}
-	public CustAlert(int custid, int typ, String tim, int usage) {
+	public CustAlert(int custid, int typ, String tim, long usage) {
 		super();
 		this.custid = custid;
 		this.typ = typ;
@@ -37,10 +37,10 @@ public class CustAlert {
 	public void setTim(String tim) {
 		this.tim = tim;
 	}
-	public int getUsage() {
+	public long getUsage() {
 		return usage;
 	}
-	public void setUsage(int usage) {
+	public void setUsage(long usage) {
 		this.usage = usage;
 	}
